+++ pkg/gatewayserver/io/mqtt/mqtt_test.go
@@ -392,7 +392,7 @@ func TestTraffic(t *testing.T) {
 					t.FailNow()
 				}
 
-				_, err := conn.ScheduleDown(tc.Path, tc.Message)
+				_, _, _, err := conn.ScheduleDown(tc.Path, tc.Message)
 				if err != nil && (tc.ErrorAssertion == nil || !a.So(tc.ErrorAssertion(err), should.BeTrue)) {
 					t.Fatalf("Unexpected error: %v", err)
 				}