+++ pkg/packetbrokeragent/mock/pbdataplane.go
@@ -19,8 +19,10 @@ import (
 	"crypto/tls"
 	"crypto/x509"
 
+	pbtypes "github.com/gogo/protobuf/types"
 	routingpb "go.packetbroker.org/api/routing"
 	packetbroker "go.packetbroker.org/api/v3"
+	"go.thethings.network/lorawan-stack/v3/pkg/ttnpb"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials"
 )
@@ -28,10 +30,12 @@ import (
 // PBDataPlane is a mock Packet Broker Data Plane.
 type PBDataPlane struct {
 	*grpc.Server
-	ForwarderUp     chan *packetbroker.RoutedUplinkMessage
-	ForwarderDown   chan *packetbroker.RoutedDownlinkMessage
-	HomeNetworkDown chan *packetbroker.RoutedDownlinkMessage
-	HomeNetworkUp   chan *packetbroker.RoutedUplinkMessage
+	ForwarderUp              chan *packetbroker.RoutedUplinkMessage
+	ForwarderDown            chan *packetbroker.RoutedDownlinkMessage
+	ForwarderDownStateChange chan *packetbroker.DownlinkMessageDeliveryStateChange
+	HomeNetworkDown          chan *packetbroker.RoutedDownlinkMessage
+	HomeNetworkUp            chan *packetbroker.RoutedUplinkMessage
+	HomeNetworkUpStateChange chan *packetbroker.UplinkMessageDeliveryStateChange
 }
 
 // NewPBDataPlane instantiates a new mock Packet Broker Data Plane.
@@ -45,26 +49,31 @@ func NewPBDataPlane(cert tls.Certificate, clientCAs *x509.CertPool) *PBDataPlane
 		Server: grpc.NewServer(
 			grpc.Creds(creds),
 		),
-		ForwarderUp:     make(chan *packetbroker.RoutedUplinkMessage),
-		ForwarderDown:   make(chan *packetbroker.RoutedDownlinkMessage),
-		HomeNetworkDown: make(chan *packetbroker.RoutedDownlinkMessage),
-		HomeNetworkUp:   make(chan *packetbroker.RoutedUplinkMessage),
+		ForwarderUp:              make(chan *packetbroker.RoutedUplinkMessage),
+		ForwarderDown:            make(chan *packetbroker.RoutedDownlinkMessage),
+		ForwarderDownStateChange: make(chan *packetbroker.DownlinkMessageDeliveryStateChange),
+		HomeNetworkDown:          make(chan *packetbroker.RoutedDownlinkMessage),
+		HomeNetworkUp:            make(chan *packetbroker.RoutedUplinkMessage),
+		HomeNetworkUpStateChange: make(chan *packetbroker.UplinkMessageDeliveryStateChange),
 	}
 	routingpb.RegisterForwarderDataServer(dp.Server, &routerForwarderServer{
-		upCh:   dp.ForwarderUp,
-		downCh: dp.ForwarderDown,
+		upCh:     dp.ForwarderUp,
+		downCh:   dp.ForwarderDown,
+		reportCh: dp.ForwarderDownStateChange,
 	})
 	routingpb.RegisterHomeNetworkDataServer(dp.Server, &routerHomeNetworkServer{
-		downCh: dp.HomeNetworkDown,
-		upCh:   dp.HomeNetworkUp,
+		downCh:   dp.HomeNetworkDown,
+		upCh:     dp.HomeNetworkUp,
+		reportCh: dp.HomeNetworkUpStateChange,
 	})
 	return dp
 }
 
 type routerForwarderServer struct {
 	routingpb.UnimplementedForwarderDataServer
-	upCh   chan *packetbroker.RoutedUplinkMessage
-	downCh chan *packetbroker.RoutedDownlinkMessage
+	upCh     chan *packetbroker.RoutedUplinkMessage
+	downCh   chan *packetbroker.RoutedDownlinkMessage
+	reportCh chan *packetbroker.DownlinkMessageDeliveryStateChange
 }
 
 func (s *routerForwarderServer) Publish(ctx context.Context, req *routingpb.PublishUplinkMessageRequest) (*routingpb.PublishUplinkMessageResponse, error) {
@@ -92,10 +101,20 @@ func (s *routerForwarderServer) Subscribe(req *routingpb.SubscribeForwarderReque
 	}
 }
 
+func (s *routerForwarderServer) ReportDownlinkMessageDeliveryState(ctx context.Context, req *routingpb.DownlinkMessageDeliveryStateChangeRequest) (*pbtypes.Empty, error) {
+	select {
+	case <-ctx.Done():
+		return nil, ctx.Err()
+	case s.reportCh <- req.StateChange:
+	}
+	return ttnpb.Empty, nil
+}
+
 type routerHomeNetworkServer struct {
 	routingpb.UnimplementedHomeNetworkDataServer
-	downCh chan *packetbroker.RoutedDownlinkMessage
-	upCh   chan *packetbroker.RoutedUplinkMessage
+	downCh   chan *packetbroker.RoutedDownlinkMessage
+	upCh     chan *packetbroker.RoutedUplinkMessage
+	reportCh chan *packetbroker.UplinkMessageDeliveryStateChange
 }
 
 func (s *routerHomeNetworkServer) Publish(ctx context.Context, req *routingpb.PublishDownlinkMessageRequest) (*routingpb.PublishDownlinkMessageResponse, error) {
@@ -129,4 +148,13 @@ func (s *routerHomeNetworkServer) Subscribe(req *routingpb.SubscribeHomeNetworkR
 			}
 		}
 	}
+}
+
+func (s *routerHomeNetworkServer) ReportUplinkMessageDeliveryState(ctx context.Context, req *routingpb.UplinkMessageDeliveryStateChangeRequest) (*pbtypes.Empty, error) {
+	select {
+	case <-ctx.Done():
+		return nil, ctx.Err()
+	case s.reportCh <- req.StateChange:
+	}
+	return ttnpb.Empty, nil
 }